+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Pyvaru Changelog
 
+## v0.1.1
+### Improvement:
+- Attribute/Key/Index errors that may occur in the Validator's get_rules() method are now catched in validate()
+implementation and the error properly reported in the ValidationResult
+
 ## v0.1.0
 ### Added: